pyproject.toml
[project]
name = "bugninja"
version = "0.1.0"
description = "AI-Powered Browser Automation"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Bugninja Team", email = "akos.jakub@bugninja.ai"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
keywords = ["browser", "automation", "ai", "testing", "web"]
dependencies = [
    "black==25.1.0",
    "browser-use[memory]==0.2.5",
    "celery[redis]==5.5.3",
    "click>=8.2.1",
    "cssselect==1.3.0",
    "cuid2==2.0.1",
    "faker==37.4.0",
    "fastapi==0.115.13",
    "hypercorn==0.17.3",
    "isort==6.0.1",
    "lxml==5.4.0",
    "mypy==1.16.0",
    "openai==1.84.0",
    "polyfactory==2.22.0",
    "pydantic==2.10.6",
    "pydantic-settings>=2.10.1",
    "pytest==8.4.1",
    "pytest-asyncio==1.0.0",
    "pytest-pretty>=1.3.0",
    "python-dotenv==1.1.0",
    "redis==5.2.1",
    "rich==14.0.0",
    "ruff==0.11.13",
    "typer>=0.16.0",
    "types-lxml==2025.3.30",
    "uvloop>=0.21.0",
]

[project.scripts]
bugninja = "bugninja_cli.main:cli"

[tool.setuptools.packages.find]
include = ["bugninja*", "bugninja_cli*"]
exclude = ["tests*", "logs*", "data_dir*", "screenshots*"]


[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"] # ignore unused imports in __init__.py files
# "test_*.py" = ["F401", "F403", "F811"]  # ignore unused imports in test files as well

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
extend_skip = ["__init__.py"]

[tool.mypy]
python_version = "3.13"
strict = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true


[tool.poe.tasks]


#? Development related commands
[tool.poe.tasks.run]
help   = "Start test traversal agent interaction"
shell = "uv run python main.py"

[tool.poe.tasks.test]
help   = "Run tests on the project"
shell = "uv run pytest tests"


[tool.poe.tasks.replay]
help   = "Replay (and potentially heal if necessary) the last saved traversal"
shell = "uv run python replay.py"

[tool.poe.tasks.lint]
help   = "Lints the whole project searching for inefficiencies in the code"
shell = "uv run ruff check --fix bugninja tests bugninja_cli . && uv run isort bugninja tests bugninja_cli . && uv run black bugninja tests bugninja_cli . && uv run mypy main.py bugninja tests bugninja_cli"

